gpioHardwareRevision: use OTP registers as fallback
This uses `vcgencmd otp_dump` to determine the board revision of the Pi
that we're running on. This is needed on 64-bit systems that do not have
the system/linux,revision DeviceTree node.

The system/linux,revision DT node is normally populated by the Raspberry
Pi's bootloader (start.elf) before jumping into the next payload. This
usually means that Linux can read this node to determine the board
revision. However, there are two cases where this doesn't work:

 - when booting via u-boot (eg. on NixOS). This happens because u-boot
   drops the provided DT and uses its own. It does not pass through the
   bootloader-created nodes, and likely won't because they are
   technically non-standard [1].

 - when booting via rpi-open-firmware - although I'm not sure anything
   else from pigpio really works (haven't tested it).

Optimally we would just retrieve the board revision via OTP ourselves,
or via a mailbox call, instead of shelling out to a third-party tool
that might not be installed. However, I don't think we have a mailbox
client for 64-bit systems (as /dev/vcio is unavailable there, instead
there's /dev/vchiq which we're not targeting). In the future, if this
instead becomes implemented some other way (eg. via /dev/vchiq) we
should probably make this _the_ main way of retrieving the board
revision (thus simplifying the multi-attempt codebase).

We also drive-by fix two smaller issues:
 - fclose() is called on fopen(dtnode) even if fopen returns NULL - this
   causes a crash on systems that don't have this device tree node
 - rev==0 is interpreted as valid, instead of failing early

[1] - http://u-boot.10912.n7.nabble.com/PATCH-v2-0-1-Pass-board-revision-and-serial-number-to-the-td359829.html